[CRUD] Download PerfCake and plugins instead of building them

+++ devtools-performance-core-crud-soak.sh
@@ -10,9 +10,14 @@ if [[ "$SERVER_HOST" == "localhost" ]];
 then
 	echo "Need a local server - preparing Docker containers..."
 
-	# Clean docker containers
+	# Check for docker
+	docker info >> /dev/null
+	[[ $? -ne 0 ]] && exit 1
+
+	# Clean docker containers, volumes, images
 	for i in `docker ps -a -q`; do docker rm -f $i; done
 	for i in `docker volume ls -q`; do docker volume rm $i; done
+	for i in `docker images -q`; do docker rmi $i; done
 
 	# Build the core server that will provide our test client with tokens
 
@@ -42,9 +47,18 @@ while true;
 do
    echo "Checking if the Core server is up and running ..."
    curl --silent http://$SERVER_HOST:$SERVER_PORT/api/status
-   [[ $? -eq 0 ]] && break
-   echo "The Core server is not responding, trying again after 10s."
-   sleep 10
+   if [[ $? -eq 0 ]]; then
+     response_code=`curl -i --silent http://$SERVER_HOST:$SERVER_PORT/api/status | head -n 1 | cut -d " " -f2`;
+     if [[ "$response_code" -eq "200" ]]; then
+       break;
+     else
+       echo "The Core server is not ready - responding by $response_code code.";
+     fi;
+   else
+     echo "The Core server is not responding.";
+   fi
+   echo "Trying again after 10s.";
+   sleep 10;
 done
 CORE_SERVER_STATUS=`curl --silent http://$SERVER_HOST:$SERVER_PORT/api/status | grep commit | sed -e 's,":",=,g' | sed -e 's,[{"}],,g' | sed -e 's,\,,;,g'`
 BASE_PERFREPO_TAGS="$ADDITIONAL_PERFREPO_TAGS;server=$SERVER_HOST:$SERVER_PORT;$CORE_SERVER_STATUS"
